refactor: move removeAllDebugLevels() from apex-log helpers to playwright-vscode-ext helpers
