+++ packages/playwright-vscode-ext/src/index.ts
@@ -22,6 +22,7 @@ export {
   isMacDesktop,
   isWindowsDesktop,
   validateNoCriticalErrors,
+  removeAllDebugLevels,
   ensureSecondarySideBarHidden,
   waitForExtensionsActivated,
   closeWorkspaceToEmptyWindow,
